[backend] Init Gradle project and rework packages
