+++ src/main/kotlin/gwent/vanilla/domain/Board.kt
@@ -1,3 +1,3 @@
-package main.domain
+package gwent.vanilla.domain
 
 data class Board(var rows: List<Row> = mutableListOf(), var power: Int = 0)
