+++ src/main/kotlin/gwent/vanilla/domain/Cards.kt
@@ -1,4 +1,4 @@
-package main.domain
+package gwent.vanilla.domain
 
 // ============ CREATURES ==========
