+++ src/main/kotlin/gwent/vanilla/domain/Round.kt
@@ -1,6 +1,7 @@
-package main.domain
+package gwent.vanilla.domain
 
-import main.domain.Round
+import gwent.vanilla.domain.Board
+import gwent.vanilla.domain.Player
 
 data class Round(val player1: Player, val player2: Player, val startingplayer: Player) {
     var boards: Map<Player, Board> = mapOf(Pair(player1, Board()), Pair(player2, Board()))
