+++ src/main/kotlin/gwent/vanilla/domain/Row.kt
@@ -1,4 +1,6 @@
-package main.domain
+package gwent.vanilla.domain
+
+import gwent.vanilla.domain.Spell
 
 data class Row(var cards: List<Spell> = mutableListOf()) {
     fun calculate_power() {
