+++ src/main/kotlin/gwent/vanilla/domain/Suit.kt
@@ -1,4 +1,4 @@
-package main.domain
+package gwent.vanilla.domain
 
 enum class Suit {
     Clubs, Diamonds, Spades, Hearts
