+++ src/test/kotlin/gwent/vanilla/AppTest.kt
@@ -1,0 +1,11 @@
+package gwent.vanilla
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class AppTest {
+    @Test
+    fun dummyTest() {
+        assertTrue { true }
+    }
+}
